// src/utils/types.ts
interface C {
  flags: { png: string };
  name: { common: string; nativeName: {} };
  population: string;
  fifa: string;
  cca3: string;
  cca2: string;
  region: string;
  subregion: string;
  capital: string[];
  currencies: {};
  tld: string;
  languages: {};
  borders: string[];
}
export type { C };